refactor: vendor shared alert helpers
