+++ alerts/infra/onchain-event-handler/src/vendored-source-drift.test.ts
@@ -1,0 +1,41 @@
+import { readFileSync } from "node:fs";
+import { join } from "node:path";
+import { describe, expect, it } from "vitest";
+
+// Vitest runs with cwd = this package root, both locally (`pnpm test`) and
+// in CI (working-directory). The repo root is three levels up.
+const packageRoot = process.cwd();
+const repoRoot = join(packageRoot, "..", "..", "..");
+
+const read = (path: string) => readFileSync(path, "utf8");
+
+describe("vendored shared-source files stay byte-identical", () => {
+  it("quicknode-hmac.ts matches the governance-watchdog copy", () => {
+    expect(read(join(packageRoot, "src", "quicknode-hmac.ts"))).toBe(
+      read(
+        join(
+          repoRoot,
+          "governance-watchdog",
+          "src",
+          "utils",
+          "quicknode-hmac.ts",
+        ),
+      ),
+    );
+  });
+
+  it("gcp-logger.ts matches the oncall-announcer copy", () => {
+    expect(read(join(packageRoot, "src", "gcp-logger.ts"))).toBe(
+      read(
+        join(
+          repoRoot,
+          "alerts",
+          "infra",
+          "oncall-announcer",
+          "src",
+          "gcp-logger.ts",
+        ),
+      ),
+    );
+  });
+});
