+++ docs/pr-checklists/recurring-review-patterns.md
@@ -99,7 +99,7 @@ ratchet is now fully in place.
 - **Indexer handlers must not bypass the RPC effect layer (blocking, `pnpm code-health:deps`)**: `indexer-envio/src/handlers/**` must not import directly from `rpc/` implementation files (`rpc/pool-state.ts`, `rpc/oracle-state.ts`, `rpc/biPoolManager.ts`, `rpc/breakers.ts`, etc.). Handlers must go through `rpc/effects.ts` (the Envio Effect API facade, which provides per-batch memoisation, deduplication, and rate-limiting) or the `rpc.ts` barrel (for DB helpers like `getPoolsByFeed`). Direct fetcher imports bypass effect deduplication and fire two RPC reads per event instead of one. Pre-inventory: zero violations; rule ships at `error`.
 - **Dead-code / dep hygiene (blocking, `pnpm --filter <pkg> knip`)**: every package runs `knip` in strict mode. Unused files / unlisted deps / binary entries are errors. Unused exports + types are warns — clean them when you touch the file. Peer-dep build tools (axe-core, tailwindcss, @stryker-mutator/api) go in `ignoreDependencies` with a 1-line "why" in this checklist.
 - **Complexity / size / cognitive-complexity budgets (blocking, diff-aware baseline)**: per-package thresholds for `complexity`, `max-lines-per-function`, `max-depth`, `max-params`, plus `eslint-plugin-sonarjs` (cognitive-complexity + 4 suspicious-pattern rules). Strictest on `shared-config`; loosest inside `indexer-envio/src/handlers/**`. Pre-existing violations live in each package's `eslint-baseline.json`; new violations fail `pnpm --filter <pkg> lint` via `scripts/eslint-baseline-diff.mjs`. See `docs/pr-checklists/code-health.md`.
-- **Duplication detection (advisory, `pnpm code-health:duplication`)**: `jscpd` scans `src/` across all packages on every PR (excluding tests, `indexer-envio/src/handlers/**`, route entry pages, and pure type modules — they're intentionally repetitive). The CI job is non-blocking — surfaces a comment-summary + an artifact in `reports/jscpd/`. Use the artifact to plan extract-helper refactors.
+- **Duplication detection (advisory, `pnpm code-health:duplication`)**: `jscpd` scans `src/` across all packages and standalone function roots on every PR (excluding tests, `indexer-envio/src/handlers/**`, route entry pages, pure type modules, and the intentionally vendored `quicknode-hmac.ts` / `gcp-logger.ts` copies — per-package drift tests enforce byte-identity for those). The CI job is non-blocking — surfaces a comment-summary + an artifact in `reports/jscpd/`. Use the artifact to plan extract-helper refactors.
 - **Code-health history report (advisory, `pnpm code-health:history`)**: writes `reports/code-health-history.md` with hotspots, change-coupling, ownership concentration, weekly delta. Run before large refactors so the targets are picked from data, not vibes. Weekly cron + Slack delivery in a follow-up PR.
 - **Per-package coverage floors (blocking, `pnpm --filter <pkg> test:coverage` or Aegis `test:cov`)**: coverage thresholds enforce a floor so deleting tests can't silently lower coverage below the baseline. Vitest packages keep `coverage.thresholds` in each `vitest.config.ts`; Aegis keeps Jest `coverageThreshold` in its Jest config. Floors are calibrated at `floor(measured) - 2` to absorb variance. CI and the local agent quality gate call coverage commands, not bare `test`, for all seven workspace packages with test suites. When adding significant new code, re-measure and raise the floors accordingly.
 - **Mutation score gates (advisory, `pnpm {bridge,dashboard,indexer}:mutation`)**: `.github/workflows/mutation-testing.yml` runs three per-package jobs on the weekly `schedule` cron + `workflow_dispatch` only — **not** per-PR. It is not in the `main` ruleset's required checks; per-PR mutation testing (3 runner boots per push) was removed as a CI-cost control (mutation measures test-suite strength, not per-commit regression — weekly is the right cadence). Run on demand for a branch via the GitHub "Run workflow" button. Every job keeps its inline `filter`/`decide` steps; on a non-`pull_request` trigger the mutation step always runs. `break` floors are calibrated at `floor(measured) − 2` for measurement noise: `metrics-bridge/stryker.config.mjs` `break: 84` (baseline 86.01%), `ui-dashboard/stryker.config.mjs` `break: 86` (baseline 88.81%), `indexer-envio/stryker.config.mjs` `break: 92` (baseline 94.19%). Surviving mutants must be classified as equivalent / accepted noise in `docs/mutation-testing.md` or fixed with a new test in the same PR. See `docs/pr-checklists/mutation-testing.md`.
